backend/requirements.txt
aider-chat==0.74.2
aider-install==0.1.3
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
ConfigArgParse==1.7
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
fastapi==0.115.8
filelock==3.17.0
flake8==7.1.1
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grep-ast==0.5.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.5.2
Jinja2==3.1.5
jiter==0.8.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.60.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
monotonic==1.6
multidict==6.1.0
narwhals==1.25.2
networkx==3.2.1
numpy==1.26.4
openai==1.61.1
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
posthog==3.11.0
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pypandoc==1.15
pyparsing==3.2.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
reportlab==4.3.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scipy==1.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
streamlit==1.42.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.29
uvicorn==0.34.0
uvloop==0.21.0
watchdog==4.0.2
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
zipp==3.21.0
